% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy_slack.R
\name{send_slack}
\alias{send_slack}
\title{Sending the message to slack channel}
\usage{
send_slack(message = "Analysis has been completed")
}
\arguments{
\item{message}{message send to slack}
}
\description{
\code{send_slack} send the message to slack channel
}
\examples{
# send_slack("Hello!")
}
