% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy_slack.R
\name{set_slack}
\alias{set_slack}
\title{Setting the slack information to system}
\usage{
set_slack(slack_token, slack_channel = "#general", slack_username = "R")
}
\arguments{
\item{slack_token}{slack acess token}

\item{slack_channel}{slack channel}

\item{slack_username}{slack username}
}
\description{
\code{set_slack} set the slack information to system
}
\examples{
# set_slack("GitHub access token", "#r", "Mr.R")
}
